requirements.txt
# Windows audio requirements
sounddevice>=0.4.5
numpy>=1.24.0
PyAudioWPatch>=0.2.12

# Transcription
openai-whisper>=20231117

# Server
fastapi>=0.100
uvicorn[standard]>=0.22
python-multipart>=0.0.6
websockets>=10.4
pydantic>=2.0
